Lot of things
Add auto update for Position and rotation
Add custom origin for all elements. This is the default origin for all drawing. It allows to have a different origin for a same asset with the same offset (thanks again to TabularElf)
Add a layer of inheritance based on the kind of asset an element use. Now each element is either from a Font type or a Sprite type. It was requiered in order to build the auto Update system as the calculation would differ if the asset would be a sprite (offset) or a text (alignement)
Update the demo with the elements going weeeeeeeeeeeeee (rotation)
Allow feedback to be apply to the Items of a Caroussel. That way you could use a feedback on the caroussel itself and a feedback on the active Item.
Rename __hudelement__ to __baseElement__
A lot more consistency, with a lot of the method being put directly in the BaseElement
Update the JabbaFont sprite

[REGRESSION] The shader for the GaugeBar Element is broken