Alpha v0.3 (approximatly)
From now on, it has the features I need for my game,so I'm going to use it in context and I'll update it accordingly. Nothing better than a proper use case to make it better.

Rewrite the way feedbacks are trigger.
Didn't test the custom feedback method. Should be broken
Add some built-in feedbacks (pretty sure they won't work with all elements)
built-in feedbacks use there own variable instead of the Element's variable.
Performance may not be the best.