Introducing Bib Fortuna
Almost down.
It allows to spawn a "notification" (called a Fortuna, because I'm going all the SW way XD) which is a special hud element spawned by a Jabba Element to show a value. Use case : a +2 for a Combo counter element or a- 1000pts for the Score Counter, etc.
[BUG] the first notification only is doing what it is expected to do. Seems I have a scoping/static issue ... again.
[WARNING] The extention is activated o the Jabba Counter only for now.