[WIP] Reworking Feedback & Jabba Container
Successful tests to change the way an element is added to a JabbaContainer
Successful tests to rework the feedback system based on how I did it for BibFortuna.
Feedback effect are now stored in the global space (took inspiration from how Juju doesn't it for Input) but I should be able to contain them into the new __feedbackPlayer constructor, I guess.
BibFortuna is now a full part of Jabba and not an extention anymore (was a pain in the ass to maintain)
add an enum to define the type of element.

All of those changes are WIP and not usable yet.